Sprints AD + L10N + CL — Ads, App i18n, Content Localization
Bundles three closed sprints that had accumulated uncommitted on main
(separating per-sprint is infeasible — same files carry changes from
multiple sprints and interactive staging is unavailable here).

Sprint AD — Advertising (AdMob), AD1–AD7:
- New Ads bounded context (schema `ads`): RewardedAdGrant append-only
  ledger, idempotency on AdMob SSV transactionId, daily-cap rule,
  GrantRewardedAdRewardCommand (verify → idempotency → cap → IDiamondGrant
  → ledger), GET /ads/rewarded/{callback,status}, fail-closed SSV verifier
  + dev fail-open verifier.
- Frontend: google_mobile_ads, AdsService over an SDK-free platform seam,
  interstitials (1/3 game start, 1/2 game end), rewarded → Diamond
  (Earn Diamonds), UMP consent + iOS ATT. Mobile-only, web-safe.

Sprint L10N — App UI i18n (Phase 1), L1–L8:
- flutter_localizations + intl + ARB (tr/en/de/fr/es) + generated
  AppLocalizations, context.l10n. LocaleCubit above the router; Settings
  language picker persists device-local + best-effort Player.Locale.
- All app UI strings extracted across gameplay/economy/quests/profile/
  admin/shared surfaces. Cubit/API error messages stay English (Phase 3).

Sprint CL — Content Localization (Phase 2), CL1–CL4:
- Category.Language (Games domain + EF + DbUp, default tr-TR, indexed),
  projected through DTOs; GET /categories?locale=xx-XX filtering.
- Language-aware CategoryImporter (language in stable ids); first en-US
  graph (docs/category-animals-en.json).
- Admin content endpoints + Flutter admin_content feature (/admin/content)
  with locale filter and create/edit language selector.
- CL4: docs/CONTENT_AUTHORING.md content-ops authoring handoff.

+++ frontend/lib/features/splash/presentation/splash_screen.dart
@@ -25,8 +25,7 @@ class _SplashScreenState extends State<SplashScreen>
   @override
   void initState() {
     super.initState();
-    _totalDuration =
-        _letterStagger * (_logoText.length - 1) + _letterDuration;
+    _totalDuration = _letterStagger * (_logoText.length - 1) + _letterDuration;
     _controller = AnimationController(vsync: this, duration: _totalDuration)
       ..addStatusListener(_onStatus)
       ..forward();
@@ -149,7 +148,7 @@ class _SandLetter extends StatelessWidget {
   Widget build(BuildContext context) {
     final eased = Curves.easeOutCubic.transform(progress);
     final dy = (1 - eased) * -_fallDistance;
-    final opacity = math.min(progress * 3, 1.0).clamp(0.0, 1.0);
+    final opacity = math.min(progress * 3, 1).clamp(0, 1).toDouble();
     final hasStarted = progress > 0 && progress < 1;
 
     return SizedBox(
